pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["aivol"]

[project]
name = "aivol"
version = "0.0.3"
authors = [{ name = "Dmitry Galchinsky", email = "galchinsky@borisfx.com" }]
description = "Utilities for AI volumes"
readme = "README.md"
license = { text = "MIT License" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=5.4",
    "pandas>=1.1"
]
